feat: Mail Send, Mail Get 생성

+++ APIServer/Model/DTO/Login.cs
@@ -3,7 +3,7 @@ using System.ComponentModel.DataAnnotations;
 public class LoginReq
 {
     [Required(ErrorMessage = "Id is required")]
-    public int UserId { get; set; }
+    public Int64 UserId { get; set; }
     
     [Required(ErrorMessage = "Token is required")]
     public string Token { get; set; } = null!;
