+++ APIServer/Model/DTO/VerifyLogin.cs
@@ -1,8 +1,6 @@
-using System.ComponentModel.DataAnnotations;
-
 public class VerifyLoginReq
 {
-    public int UserId { get; set; }
+    public Int64 UserId { get; set; }
     public string Token { get; set; } = null!;
 }
