+++ HiveServer/Model/DAO/UserData.cs
@@ -1,6 +1,6 @@
 public class UserData
 {
-    public int user_id { get; set; }
+    public Int64 user_id { get; set; }
     public string password { get; set; } = null!;
     public string email { get; set; } = null!;
     public DateTime? created_at { get; set; }
